pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "moge"
version = "2.0.0"
description = "MoGe: Unlocking Accurate Monocular Geometry Estimation for Open-Domain Images with Optimal Training Supervision"
readme = "README.md"
license = {text = "MIT"}
dependencies = [
    "click",
    "opencv-python",
    "scipy",
    "matplotlib",
    "trimesh",
    "pillow",
    "huggingface_hub",
    "numpy",
    "torch>=2.0.0",
    "torchvision",
    "gradio",
    "utils3d @ git+https://github.com/EasternJournalist/utils3d.git@3fab839f0be9931dac7c8488eb0e1600c236e183",
    "pipeline @ git+https://github.com/EasternJournalist/pipeline.git@866f059d2a05cde05e4a52211ec5051fd5f276d6"
]
requires-python = ">=3.9"

[project.urls]
Homepage = "https://github.com/microsoft/MoGe"

[tool.setuptools.packages.find]
where = ["."]
include = ["moge*"]

[project.scripts]
moge = "moge.scripts.cli:main"